ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai==0.28.1
PyPDF2==3.0.1
python-docx==1.1.0
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0
python-dotenv==1.0.0
flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.1.78
numpy==1.24.3
ultralytics==8.0.207
Pillow==10.0.1
face-recognition==1.3.0